Update simplify-fractions.py
This version fixes the issue with the numerator and denominator being outputed as decimals. Could have also used // to accomplish it.

+++ simplify-fractions.py
@@ -10,4 +10,4 @@ def find_lowest(num, denom):
 
 for x in range (0, len(numbers),2):
     lowest = find_lowest(int(numbers[x]), int(numbers[x+1]))
-    print(numbers[x] + "/" + numbers[x+1] + " simplifies to: " + str(int(numbers[x]) / lowest) + "/" + str(int(numbers[x+1])/lowest))
+    print(numbers[x] + "/" + numbers[x+1] + " simplifies to: " + str(int(int(numbers[x]) / lowest)) + "/" + str(int((int(numbers[x+1]) / lowest))))
